test: refactored tests and code
